Some renaming and adding more comments
